# pb/role_message.pb.go
//

// Code generated by protoc-gen-go. DO NOT EDIT.
// versions:
// 	protoc-gen-go v1.26.0
// 	protoc        v3.6.1
// source: role_message.proto

package pb

import (
	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
	reflect "reflect"
	sync "sync"
)

const (
	// Verify that this generated code is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
	// Verify that runtime/protoimpl is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
)

// Permissions
type Permissions int32

const (
	//
	Permissions_UNKNOWN Permissions = 0
	//
	Permissions_READWRITE Permissions = 1
	//
	Permissions_READ Permissions = 2
	//
	Permissions_WRITE Permissions = 3
)

// Enum value maps for Permissions.
var (
	Permissions_name = map[int32]string{
		0: "UNKNOWN",
		1: "READWRITE",
		2: "READ",
		3: "WRITE",
	}
	Permissions_value = map[string]int32{
		"UNKNOWN":   0,
		"READWRITE": 1,
		"READ":      2,
		"WRITE":     3,
	}
)

func (x Permissions) Enum() *Permissions {
	p := new(Permissions)
	*p = x
	return p
}

func (x Permissions) String() string {
	return protoimpl.X.EnumStringOf(x.Descriptor(), protoreflect.EnumNumber(x))
}

func (Permissions) Descriptor() protoreflect.EnumDescriptor {
	return file_role_message_proto_enumTypes[0].Descriptor()
}

func (Permissions) Type() protoreflect.EnumType {
	return &file_role_message_proto_enumTypes[0]
}

func (x Permissions) Number() protoreflect.EnumNumber {
	return protoreflect.EnumNumber(x)
}

// Deprecated: Use Permissions.Descriptor instead.
func (Permissions) EnumDescriptor() ([]byte, []int) {
	return file_role_message_proto_rawDescGZIP(), []int{0}
}

// Name
type Name int32

const (
	//
	Name_UNKNOWN_NAME Name = 0
	//
	Name_OWNER Name = 1
	//
	Name_ADMIN Name = 2
	//
	Name_CUSTOMER Name = 3
)

// Enum value maps for Name.
var (
	Name_name = map[int32]string{
		0: "UNKNOWN_NAME",
		1: "OWNER",
		2: "ADMIN",
		3: "CUSTOMER",
	}
	Name_value = map[string]int32{
		"UNKNOWN_NAME": 0,
		"OWNER":        1,
		"ADMIN":        2,
		"CUSTOMER":     3,
	}
)

func (x Name) Enum() *Name {
	p := new(Name)
	*p = x
	return p
}

func (x Name) String() string {
	return protoimpl.X.EnumStringOf(x.Descriptor(), protoreflect.EnumNumber(x))
}

func (Name) Descriptor() protoreflect.EnumDescriptor {
	return file_role_message_proto_enumTypes[1].Descriptor()
}

func (Name) Type() protoreflect.EnumType {
	return &file_role_message_proto_enumTypes[1]
}

func (x Name) Number() protoreflect.EnumNumber {
	return protoreflect.EnumNumber(x)
}

// Deprecated: Use Name.Descriptor instead.
func (Name) EnumDescriptor() ([]byte, []int) {
	return file_role_message_proto_rawDescGZIP(), []int{1}
}

// Role
type Role struct {
	state         protoimpl.MessageState
	sizeCache     protoimpl.SizeCache
	unknownFields protoimpl.UnknownFields

	// Role Name
	Name Name `protobuf:"varint,1,opt,name=name,proto3,enum=Name" json:"name,omitempty"`
	// Permissions
	Permissions []Permissions `protobuf:"varint,2,rep,packed,name=permissions,proto3,enum=Permissions" json:"permissions,omitempty"`
}

func (x *Role) Reset() {
	*x = Role{}
	if protoimpl.UnsafeEnabled {
		mi := &file_role_message_proto_msgTypes[0]
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		ms.StoreMessageInfo(mi)
	}
}

func (x *Role) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*Role) ProtoMessage() {}

func (x *Role) ProtoReflect() protoreflect.Message {
	mi := &file_role_message_proto_msgTypes[0]
	if protoimpl.UnsafeEnabled && x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use Role.ProtoReflect.Descriptor instead.
func (*Role) Descriptor() ([]byte, []int) {
	return file_role_message_proto_rawDescGZIP(), []int{0}
}

func (x *Role) GetName() Name {
	if x != nil {
		return x.Name
	}
	return Name_UNKNOWN_NAME
}

func (x *Role) GetPermissions() []Permissions {
	if x != nil {
		return x.Permissions
	}
	return nil
}

var File_role_message_proto protoreflect.FileDescriptor

var file_role_message_proto_rawDesc = []byte{
	0x0a, 0x12, 0x72, 0x6f, 0x6c, 0x65, 0x5f, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x70,
	0x72, 0x6f, 0x74, 0x6f, 0x22, 0x51, 0x0a, 0x04, 0x52, 0x6f, 0x6c, 0x65, 0x12, 0x19, 0x0a, 0x04,
	0x6e, 0x61, 0x6d, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0e, 0x32, 0x05, 0x2e, 0x4e, 0x61, 0x6d,
	0x65, 0x52, 0x04, 0x6e, 0x61, 0x6d, 0x65, 0x12, 0x2e, 0x0a, 0x0b, 0x70, 0x65, 0x72, 0x6d, 0x69,
	0x73, 0x73, 0x69, 0x6f, 0x6e, 0x73, 0x18, 0x02, 0x20, 0x03, 0x28, 0x0e, 0x32, 0x0c, 0x2e, 0x50,
	0x65, 0x72, 0x6d, 0x69, 0x73, 0x73, 0x69, 0x6f, 0x6e, 0x73, 0x52, 0x0b, 0x70, 0x65, 0x72, 0x6d,
	0x69, 0x73, 0x73, 0x69, 0x6f, 0x6e, 0x73, 0x2a, 0x3e, 0x0a, 0x0b, 0x50, 0x65, 0x72, 0x6d, 0x69,
	0x73, 0x73, 0x69, 0x6f, 0x6e, 0x73, 0x12, 0x0b, 0x0a, 0x07, 0x55, 0x4e, 0x4b, 0x4e, 0x4f, 0x57,
	0x4e, 0x10, 0x00, 0x12, 0x0d, 0x0a, 0x09, 0x52, 0x45, 0x41, 0x44, 0x57, 0x52, 0x49, 0x54, 0x45,
	0x10, 0x01, 0x12, 0x08, 0x0a, 0x04, 0x52, 0x45, 0x41, 0x44, 0x10, 0x02, 0x12, 0x09, 0x0a, 0x05,
	0x57, 0x52, 0x49, 0x54, 0x45, 0x10, 0x03, 0x2a, 0x3c, 0x0a, 0x04, 0x4e, 0x61, 0x6d, 0x65, 0x12,
	0x10, 0x0a, 0x0c, 0x55, 0x4e, 0x4b, 0x4e, 0x4f, 0x57, 0x4e, 0x5f, 0x4e, 0x41, 0x4d, 0x45, 0x10,
	0x00, 0x12, 0x09, 0x0a, 0x05, 0x4f, 0x57, 0x4e, 0x45, 0x52, 0x10, 0x01, 0x12, 0x09, 0x0a, 0x05,
	0x41, 0x44, 0x4d, 0x49, 0x4e, 0x10, 0x02, 0x12, 0x0c, 0x0a, 0x08, 0x43, 0x55, 0x53, 0x54, 0x4f,
	0x4d, 0x45, 0x52, 0x10, 0x03, 0x42, 0x05, 0x5a, 0x03, 0x2f, 0x70, 0x62, 0x62, 0x06, 0x70, 0x72,
	0x6f, 0x74, 0x6f, 0x33,
}

var (
	file_role_message_proto_rawDescOnce sync.Once
	file_role_message_proto_rawDescData = file_role_message_proto_rawDesc
)

func file_role_message_proto_rawDescGZIP() []byte {
	file_role_message_proto_rawDescOnce.Do(func() {
		file_role_message_proto_rawDescData = protoimpl.X.CompressGZIP(file_role_message_proto_rawDescData)
	})
	return file_role_message_proto_rawDescData
}

var file_role_message_proto_enumTypes = make([]protoimpl.EnumInfo, 2)
var file_role_message_proto_msgTypes = make([]protoimpl.MessageInfo, 1)
var file_role_message_proto_goTypes = []interface{}{
	(Permissions)(0), // 0: Permissions
	(Name)(0),        // 1: Name
	(*Role)(nil),     // 2: Role
}
var file_role_message_proto_depIdxs = []int32{
	1, // 0: Role.name:type_name -> Name
	0, // 1: Role.permissions:type_name -> Permissions
	2, // [2:2] is the sub-list for method output_type
	2, // [2:2] is the sub-list for method input_type
	2, // [2:2] is the sub-list for extension type_name
	2, // [2:2] is the sub-list for extension extendee
	0, // [0:2] is the sub-list for field type_name
}

func init() { file_role_message_proto_init() }
func file_role_message_proto_init() {
	if File_role_message_proto != nil {
		return
	}
	if !protoimpl.UnsafeEnabled {
		file_role_message_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
			switch v := v.(*Role); i {
			case 0:
				return &v.state
			case 1:
				return &v.sizeCache
			case 2:
				return &v.unknownFields
			default:
				return nil
			}
		}
	}
	type x struct{}
	out := protoimpl.TypeBuilder{
		File: protoimpl.DescBuilder{
			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
			RawDescriptor: file_role_message_proto_rawDesc,
			NumEnums:      2,
			NumMessages:   1,
			NumExtensions: 0,
			NumServices:   0,
		},
		GoTypes:           file_role_message_proto_goTypes,
		DependencyIndexes: file_role_message_proto_depIdxs,
		EnumInfos:         file_role_message_proto_enumTypes,
		MessageInfos:      file_role_message_proto_msgTypes,
	}.Build()
	File_role_message_proto = out.File
	file_role_message_proto_rawDesc = nil
	file_role_message_proto_goTypes = nil
	file_role_message_proto_depIdxs = nil
}